# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Wormhole Guardian API\nThis is the API for the Wormhole Guardian and Explorer.\nThe API has two namespaces: wormholescan and guardian.\nwormholescan is the namespace for the explorer and the new endpoints. The prefix is /api/v1.\nguardian is the legacy namespace backguard compatible with guardian node API. The prefix is /v1.\nThis API is public and does not require authentication although some endpoints are rate limited.\nCheck each endpoint documentation for more information.",
        "title": "Wormholescan API",
        "termsOfService": "https://wormhole.com/",
        "contact": {
            "name": "API Support",
            "url": "https://discord.com/invite/wormholecrypto",
            "email": "info@wormhole.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/address/{address}": {
            "get": {
                "description": "Lookup an address",
                "tags": [
                    "wormholescan"
                ],
                "operationId": "find-address-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number. Starts at 0.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of elements per page.",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response-address_AddressOverview"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/v1/relays/:chain/:emitter/:sequence": {
            "get": {
                "description": "Get a specific relay information by chainID, emitter address and sequence.",
                "tags": [
                    "wormholescan"
                ],
                "operationId": "find-relay-by-vaa-id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/relays.RelayResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/swagger.json": {
            "get": {
                "description": "Returns the swagger specification for this API.",
                "tags": [
                    "wormholescan"
                ],
                "operationId": "swagger",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "address.AddressOverview": {
            "type": "object",
            "properties": {
                "vaas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vaa.VaaDoc"
                    }
                }
            }
        },
...
        "relays.DeliveryReponse": {
            "type": "object",
            "properties": {
                "execution": {
                    "$ref": "#/definitions/relays.ResultExecutionResponse"
                },
                "relayGasUsed": {
                    "type": "integer"
                }
            }
        },
        "relays.InstructionsResponse": {
            "type": "object",
            "properties": {
                "encodedExecutionInfo": {
                    "type": "string"
                },
                "extraReceiverValue": {
                    "type": "object",
                    "properties": {
                        "_hex": {
                            "type": "string"
                        },
                        "_isBigNumber": {
                            "type": "boolean"
                        }
                    }
                },
                "refundAddress": {
                    "type": "string"
                },
                "refundChainId": {
                    "type": "integer"
                },
                "refundDeliveryProvider": {
                    "type": "string"
                },
                "requestedReceiverValue": {
                    "type": "object",
                    "properties": {
                        "_hex": {
                            "type": "string"
                        },
                        "_isBigNumber": {
                            "type": "boolean"
                        }
                    }
                },
                "senderAddress": {
                    "type": "string"
                },
                "sourceDeliveryProvider": {
                    "type": "string"
                },
                "targetAddress": {
                    "type": "string"
                },
                "targetChainId": {
                    "type": "integer"
                },
                "vaaKeys": {
                    "type": "array",
                    "items": {}
                }
            }
        },
        "relays.RelayDataResponse": {
            "type": "object",
            "properties": {
                "delivery": {
                    "$ref": "#/definitions/relays.DeliveryReponse"
                },
                "fromTxHash": {
                    "type": "string"
                },
                "instructions": {
                    "$ref": "#/definitions/relays.InstructionsResponse"
                },
                "toTxHash": {
                    "type": "string"
                }
            }
        },
        "relays.RelayResponse": {
            "type": "object",
            "properties": {
                "completedAt": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/relays.RelayDataResponse"
                },
                "failedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "receivedAt": {
                    "type": "string"
                },
                "relayer": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "relays.ResultExecutionResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "gasUsed": {
                    "type": "string"
                },
                "refundStatus": {
                    "type": "string"
                },
                "revertString": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "transactionHash": {
                    "type": "string"
                }
            }
        },
        "response.Response-address_AddressOverview": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/address.AddressOverview"
                },
                "pagination": {
                    "$ref": "#/definitions/response.ResponsePagination"
                }
            }
        },
...
        "response.ResponsePagination": {
            "type": "object",
            "properties": {
                "next": {
                    "type": "string"
                }
            }
        },
...
        "vaa.ChainID": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14,
                15,
                16,
                17,
                18,
                19,
                21,
                22,
                23,
                24,
                26,
                28,
                29,
                30,
                32,
                3104,
                10002
            ],
            "x-enum-varnames": [
                "ChainIDUnset",
                "ChainIDSolana",
                "ChainIDEthereum",
                "ChainIDTerra",
                "ChainIDBSC",
                "ChainIDPolygon",
                "ChainIDAvalanche",
                "ChainIDOasis",
                "ChainIDAlgorand",
                "ChainIDAurora",
                "ChainIDFantom",
                "ChainIDKarura",
                "ChainIDAcala",
                "ChainIDKlaytn",
                "ChainIDCelo",
                "ChainIDNear",
                "ChainIDMoonbeam",
                "ChainIDNeon",
                "ChainIDTerra2",
                "ChainIDInjective",
                "ChainIDSui",
                "ChainIDAptos",
                "ChainIDArbitrum",
                "ChainIDOptimism",
                "ChainIDPythNet",
                "ChainIDXpla",
                "ChainIDBtc",
                "ChainIDBase",
                "ChainIDSei",
                "ChainIDWormchain",
                "ChainIDSepolia"
            ]
        },
        "vaa.VaaDoc": {
            "type": "object",
            "properties": {
                "appId": {
                    "description": "AppId is an extension field - it is not present in the guardian API.",
                    "type": "string"
                },
                "emitterAddr": {
                    "type": "string"
                },
                "emitterChain": {
                    "$ref": "#/definitions/vaa.ChainID"
                },
                "emitterNativeAddr": {
                    "type": "string"
                },
                "guardianSetIndex": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "indexedAt": {
                    "type": "string"
                },
                "payload": {
                    "description": "Payload is an extension field - it is not present in the guardian API.",
                    "type": "object",
                    "additionalProperties": true
                },
                "timestamp": {
                    "type": "string"
                },
                "txHash": {
                    "description": "TxHash is an extension field - it is not present in the guardian API.",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "vaa": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
    }
}